requirements.txt
boto3==1.40.13
botocore==1.40.13
Cartopy==0.23.0
certifi==2024.6.2
cftime==1.6.4
cmcrameri==1.9
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
h5netcdf==1.3.0
h5py==3.11.0
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.9.0
netCDF4==1.7.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
progress==1.6.1
progressbar2==4.5.0
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2024.1
s3transfer==0.13.1
scipy==1.15.3
shapely==2.0.4
six==1.16.0
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.5.0
xarray==2024.6.0
